chore: proxy as a separate module (#865)
