+++ proxy/src/error.rs
@@ -1,7 +1,7 @@
 // Copyright 2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
 use ceresdbproto::common::ResponseHeader;
-use common_util::error::GenericError;
+use common_util::{define_result, error::GenericError};
 use http::StatusCode;
 use snafu::{Backtrace, Snafu};
