+++ proxy/src/grpc/sql_query.rs
@@ -30,7 +30,7 @@ use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
 use tonic::{transport::Channel, IntoRequest};
 
-use crate::proxy::{
+use crate::{
     error::{self, ErrNoCause, ErrWithCause, Error, Result},
     forward::{ForwardRequest, ForwardResult},
     Context, Proxy,
