+++ proxy/src/handlers/query.rs
