+++ proxy/src/http/mod.rs
@@ -1,4 +1,4 @@
 // Copyright 2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
-pub(crate) mod prom;
-pub(crate) mod query;
+pub mod prom;
+pub mod query;
