+++ proxy/src/schema_config_provider/mod.rs
@@ -1,10 +1,11 @@
-// Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
+// Copyright 2022-2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
 //! Schema configuration can be retrieved from the the [`SchemaConfigProvider`].
 
 use std::sync::Arc;
 
 use cluster::config::SchemaConfig;
+use common_util::define_result;
 use snafu::Snafu;
 
 pub mod cluster_based;
