+++ server/src/mysql/service.rs
@@ -5,15 +5,13 @@ use std::{net::SocketAddr, sync::Arc, time::Duration};
 use common_util::runtime::JoinHandle;
 use log::{error, info};
 use opensrv_mysql::AsyncMysqlIntermediary;
+use proxy::instance::{Instance, InstanceRef};
 use query_engine::executor::Executor as QueryExecutor;
 use router::RouterRef;
 use table_engine::engine::EngineRuntimes;
 use tokio::sync::oneshot::{self, Receiver, Sender};
 
-use crate::{
-    instance::{Instance, InstanceRef},
-    mysql::{error::Result, worker::MysqlWorker},
-};
+use crate::mysql::{error::Result, worker::MysqlWorker};
 
 pub struct MysqlService<Q> {
     instance: InstanceRef<Q>,
